Transformers: Refactor single loop transformation
