+++ src/CMakeLists.txt
@@ -57,6 +57,7 @@ target_sources(golem_lib
     PRIVATE transformers/FalseClauseRemoval.cc
     PRIVATE transformers/RemoveUnreachableNodes.cc
     PRIVATE transformers/TransformationPipeline.cc
+    PRIVATE transformers/SingleLoopTransformation.cc
     )
 
 target_include_directories(golem_lib PUBLIC ${CMAKE_SOURCE_DIR}/src ${CMAKE_SOURCE_DIR}/src/include)
