+++ src/engine/Common.cc
@@ -1,5 +1,7 @@
 #include "Common.h"
 
+#include "transformers/SingleLoopTransformation.h"
+
 VerificationResult solveTrivial(ChcDirectedGraph const & graph, Logic & logic) {
     // All edges should be between entry and exit, check if any of them has a satisfiable label
     auto edgeIds = graph.getEdges();
